第二十九次上传

+++ src/main/java/com/sky/L662.java
@@ -28,13 +28,13 @@ public class L662 {
     // 广度优先搜索
     public int widthOfBinaryTree(TreeNode root) {
         int res = 1;
-        List<Pair<TreeNode, Integer>> arr = new ArrayList<>();
-        arr.add(new Pair<>(root, 1));
+        List<Pair<TreeNode, Long>> arr = new ArrayList<>();
+        arr.add(new Pair<>(root, 1L));
         while (!arr.isEmpty()) {
-            List<Pair<TreeNode, Integer>> tmp = new ArrayList<>();
-            for (Pair<TreeNode, Integer> pair : arr) {
+            List<Pair<TreeNode, Long>> tmp = new ArrayList<>();
+            for (Pair<TreeNode, Long> pair : arr) {
                 TreeNode node = pair.getKey();
-                int index = pair.getValue();
+                long index = pair.getValue();
                 if (node.left != null) {
                     tmp.add(new Pair<>(node.left, index * 2));
                 }
@@ -42,7 +42,7 @@ public class L662 {
                     tmp.add(new Pair<>(node.right, index * 2 + 1));
                 }
             }
-            res = Math.max(res, arr.get(arr.size() - 1).getValue() - arr.get(0).getValue() + 1);
+            res = Math.toIntExact(Math.max(res, arr.get(arr.size() - 1).getValue() - arr.get(0).getValue() + 1));
             arr = tmp;
         }
         return res;
